# ConfigItem: Fix infinite recursion caused by `ignore_on_error` when …