@julianbrost 2022-06-20
Please update the PR description to specify the config that was used in each case and clearly state when a process did not terminate when it should or when it crashed.

Also you have a typo in the PR title and commit message, you might want to fix that as well.